Created a folder for GPUMemory
The reason for why 32 files has changed is because I moved two classes into a new folder. So I had to change #includes on a few places in the code.
