+++ StortSpelProjekt/Engine/src/ECS/SceneManager.cpp
@@ -8,7 +8,7 @@
 // Renderer
 #include "../Renderer/CommandInterface.h"
 #include "../Renderer/ShaderResourceView.h"
-#include "../Renderer/ConstantBuffer.h"
+#include "../Renderer/GPUMemory/ConstantBuffer.h"
 #include "../Renderer/ShadowInfo.h"
 #include "../Renderer/ViewPool.h"
 #include "../Renderer/Texture.h"
