+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/BlendRenderTask.cpp
@@ -7,7 +7,7 @@
 #include "../RenderTarget.h"
 #include "../DescriptorHeap.h"
 #include "../SwapChain.h"
-#include "../Resource.h"
+#include "../GPUMemory/Resource.h"
 #include "../PipelineState.h"
 #include "../Renderer/Transform.h"
 #include "../Renderer/Mesh.h"
