+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/CopyOnDemandTask.cpp
@@ -2,7 +2,7 @@
 #include "CopyOnDemandTask.h"
 
 #include "../Texture.h"
-#include "../Resource.h"
+#include "../GPUMemory/Resource.h"
 #include "../CommandInterface.h"
 
 CopyOnDemandTask::CopyOnDemandTask(ID3D12Device5* device)
