+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/DX12Task.cpp
@@ -3,7 +3,7 @@
 
 #include "../CommandInterface.h"
 #include "../DescriptorHeap.h"
-#include "../Resource.h"
+#include "../GPUMemory/Resource.h"
 
 DX12Task::DX12Task(ID3D12Device5* device, COMMAND_INTERFACE_TYPE interfaceType)
 {
