+++ StortSpelProjekt/Engine/src/Renderer/GPUMemory/Resource.h
