+++ StortSpelProjekt/Engine/src/Renderer/RenderTarget.cpp
@@ -2,7 +2,7 @@
 #include "RenderTarget.h"
 
 #include "DescriptorHeap.h"
-#include "Resource.h"
+#include "GPUMemory/Resource.h"
 #include "RenderView.h"
 
 RenderTarget::RenderTarget(
